added tests for validator

+++ src/main/java/com/ua/foxminded/task_11/services/FacultyServices.java
@@ -1,20 +1,21 @@
 package com.ua.foxminded.task_11.services;
 
 import com.ua.foxminded.task_11.dao.impl.FacultyDaoImpl;
-import com.ua.foxminded.task_11.exceptions.DaoException;
 import com.ua.foxminded.task_11.exceptions.ServicesException;
 import com.ua.foxminded.task_11.model.Faculty;
-
 import com.ua.foxminded.task_11.validation.ValidatorEntity;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Service;
 
 import javax.validation.*;
 import java.util.List;
 import java.util.Set;
 
 
-@Component
+@Service
 public class FacultyServices {
 
     @Autowired
@@ -27,38 +28,44 @@ public class FacultyServices {
     public List<Faculty> getFaculties() {
         try {
             return facultyDao.getAll();
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to get faculties", e);
+        } catch (DataAccessException e) {
+            throw new ServicesException("Failed to get list of faculties", e);
         }
     }
 
     public boolean createNewFaculty(@Valid Faculty faculty) {
         try {
             return facultyDao.create(faculty);
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to create faculty", e);
+        } catch (DataAccessException e) {
+            throw new ServicesException("Failed to create new faculty", e);
         }
     }
 
     public boolean deleteFaculty(long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         try {
             return facultyDao.delete(id);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to delete faculty by id", e);
         }
     }
 
-    public Faculty getFaculty(long id) throws ServicesException {
+    public Faculty getFaculty(long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         Faculty faculty;
         try {
             faculty = facultyDao.getById(id);
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to retrieve faculty by id", e);
+        } catch (DataAccessException e) {
+            throw new ServicesException("Failed to retrieve faculty with such id", e);
         }
         return faculty;
     }
 
-    public boolean updateFaculty(Faculty faculty) throws ServicesException {
+    public boolean updateFaculty(@Valid Faculty faculty) {
         if (faculty.getFacultyId() == 0) {
             throw new ServicesException("Missing id");
         }
@@ -70,13 +77,13 @@ public class FacultyServices {
 
         try {
             facultyDao.getById(faculty.getFacultyId());
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to retrieve faculty by id", e);
+        } catch (DataAccessException e) {
+            throw new ServicesException("Failed to retrieve faculty with such id: ", e);
         }
 
         try {
             return facultyDao.update(faculty);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Problem with updating faculty");
         }
     }
