+++ src/main/java/com/ua/foxminded/task_11/services/GroupServices.java
@@ -1,21 +1,20 @@
 package com.ua.foxminded.task_11.services;
 
 import com.ua.foxminded.task_11.dao.impl.GroupDaoImpl;
-import com.ua.foxminded.task_11.exceptions.DaoException;
 import com.ua.foxminded.task_11.exceptions.ServicesException;
-import com.ua.foxminded.task_11.model.Faculty;
 import com.ua.foxminded.task_11.model.Group;
 import com.ua.foxminded.task_11.validation.ValidatorEntity;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Service;
 
 
 import javax.validation.*;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
-@Component
+@Service
 public class GroupServices {
     @Autowired
     private GroupDaoImpl groupDao;
@@ -23,41 +22,47 @@ public class GroupServices {
     @Autowired
     private ValidatorEntity validator;
 
-    public List<Group> getGroups() throws ServicesException {
+    public List<Group> getGroups()  {
         try {
             return groupDao.getAll();
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to get groups", e);
         }
     }
 
-    public boolean createNewGroup(@Valid Group group) throws ServicesException {
+    public boolean createNewGroup(@Valid Group group) {
         try {
             return groupDao.create(group);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to create group", e);
         }
     }
 
-    public boolean deleteGroup(long id) throws ServicesException {
+    public boolean deleteGroup(long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         try {
             return groupDao.delete(id);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to delete group by id", e);
         }
     }
 
-    public Group getGroup(long id) throws ServicesException {
+    public Group getGroup(long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         Group group;
         try {
             group = groupDao.getById(id);
-        } catch (DaoException e) {
+        } catch (DataAccessException e){
             throw new ServicesException("Failed to retrieve group by id", e);
         }
         return group;
     }
 
-    public boolean updateGroup(@Valid Group group) throws ServicesException {
+    public boolean updateGroup(@Valid Group group) {
         if (group.getGroupId() == 0) {
             throw new ServicesException("Missing id");
         }
@@ -69,22 +74,25 @@ public class GroupServices {
 
         try {
             groupDao.getById(group.getGroupId());
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to retrieve faculty from id", e);
+        } catch (DataAccessException e){
+            throw new ServicesException("Failed to retrieve faculty from such id: ", e);
         }
 
         try {
             return groupDao.update(group);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Problem with updating faculty");
         }
     }
 
-    public int getLessonsForGroup(Long id) throws ServicesException {
+    public int getLessonsForGroup(Long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         try {
             return groupDao.getLessonsById(id);
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to get lessons group by id", e);
+        } catch (DataAccessException e) {
+            throw new ServicesException("Failed to get lessons group by such id", e);
         }
 
     }
