+++ src/main/java/com/ua/foxminded/task_11/services/TimeSlotServices.java
@@ -1,14 +1,15 @@
 package com.ua.foxminded.task_11.services;
 
 import com.ua.foxminded.task_11.dao.impl.TimeSlotDaoImpl;
-import com.ua.foxminded.task_11.exceptions.DaoException;
 import com.ua.foxminded.task_11.exceptions.ServicesException;
 
 import com.ua.foxminded.task_11.model.*;
 import com.ua.foxminded.task_11.validation.ValidatorEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Service;
 
 import javax.validation.*;
 
@@ -16,7 +17,7 @@ import java.util.List;
 import java.util.Set;
 
 
-@Component
+@Service
 @Scope(value = "session")
 public class TimeSlotServices {
 
@@ -26,42 +27,48 @@ public class TimeSlotServices {
     @Autowired
     private ValidatorEntity validator;
 
-    public List<TimeSlot> getTimeSlots() throws ServicesException {
+    public List<TimeSlot> getTimeSlots() {
         try {
             return timeSlotDao.getAll();
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to get time slots", e);
         }
     }
 
 
-    public boolean createNewTimeSlot(@Valid TimeSlot timeSlot) throws ServicesException {
+    public boolean createNewTimeSlot(@Valid TimeSlot timeSlot) {
         try {
             return timeSlotDao.create(timeSlot);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to create time slot", e);
         }
     }
 
-    public boolean deleteTimeSlot(long id) throws ServicesException {
+    public boolean deleteTimeSlot(long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         try {
             return timeSlotDao.delete(id);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Failed to delete time slot by id", e);
         }
     }
 
-    public TimeSlot getTimeSlot(long id) throws ServicesException {
+    public TimeSlot getTimeSlot(long id) {
+        if (id == 0) {
+            throw new ServicesException("Missing id");
+        }
         TimeSlot timeSlot;
         try {
             timeSlot = timeSlotDao.getById(id);
-        } catch (DaoException e) {
+        } catch (EmptyResultDataAccessException e) {
             throw new ServicesException("Failed to retrieve time slot by id", e);
         }
         return timeSlot;
     }
 
-    public boolean updateTimeSlot(@Valid TimeSlot timeSlot) throws ServicesException {
+    public boolean updateTimeSlot(@Valid TimeSlot timeSlot) {
         if (timeSlot.getTimeSlotId() == 0) {
             throw new ServicesException("Missing id");
         }
@@ -73,13 +80,13 @@ public class TimeSlotServices {
 
         try {
             timeSlotDao.getById(timeSlot.getTimeSlotId());
-        } catch (DaoException e) {
-            throw new ServicesException("Failed to retrieve lesson by id", e);
+        } catch (DataAccessException e) {
+            throw new ServicesException("Failed to retrieve lesson by such id:", e);
         }
 
         try {
             return timeSlotDao.update(timeSlot);
-        } catch (DaoException e) {
+        } catch (DataAccessException e) {
             throw new ServicesException("Problem with updating lesson");
         }
     }
